+++ tests/tada_packages_tests.adb
@@ -7,14 +7,14 @@ package body Tada_Packages_Tests is
    use Ada;
    use Tada;
 
-   package Package_Names_To_Valid is new Containers.Indefinite_Hashed_Maps
+   package Strings_To_Valid is new Containers.Indefinite_Hashed_Maps
      (Key_Type => String,
       Element_Type => Boolean,
       Hash => Strings.Hash,
       Equivalent_Keys => "=");
 
    procedure Test_Validate_Package_Names (T : in out Test_Context) is
-      Names : constant Package_Names_To_Valid.Map :=
+      Names : constant Strings_To_Valid.Map :=
         ["hello" => True,
          "Hello" => False,
          "hello_world" => True,
@@ -35,12 +35,47 @@ package body Tada_Packages_Tests is
    begin
       for C in Names.Iterate loop
          declare
-            Package_Name : constant String := Package_Names_To_Valid.Key (C);
-            Validity : constant Boolean := Package_Names_To_Valid.Element (C);
+            Package_Name : constant String := Strings_To_Valid.Key (C);
+            Validity : constant Boolean := Strings_To_Valid.Element (C);
          begin
             T.Expect (Packages.Is_Valid_Name (Package_Name) = Validity,
                       "Expected: '" & Package_Name & "' to be: " & Validity'Image);
          end;
       end loop;
    end Test_Validate_Package_Names;
+
+   procedure Test_Validate_Package_Versions (T : in out Test_Context) is
+      Versions : constant Strings_To_Valid.Map :=
+        ["" => False,
+         "abc" => False,
+         "..." => False,
+         "1" => False,
+         "1.1" => False,
+         "1.1.1" => True,
+         "0.0.0" => True,
+         "01.0.0" => False,
+         "0.01.0" => False,
+         "0.0.01" => False,
+         "1.1.1.1" => False,
+         "v1.0.0" => False,
+         "1.0.a" => False,
+         "1.0.0a" => False,
+         "1.1.1-dev" => True,
+         "1.0.0-rc2" => True,
+         "1.0.0-beta_1" => True,
+         "1.0.0-123" => True,
+         "1.0.0-" => False,
+         "1.0.0-a-b" => False,
+         "1.1-foo" => False];
+   begin
+      for C in Versions.Iterate loop
+         declare
+            Package_Version : constant String := Strings_To_Valid.Key (C);
+            Validity : constant Boolean := Strings_To_Valid.Element (C);
+         begin
+            T.Expect (Packages.Is_Valid_Version (Package_Version) = Validity,
+                      "Expected: '" & Package_Version & "' to be: " & Validity'Image);
+         end;
+      end loop;
+   end Test_Validate_Package_Versions;
 end Tada_Packages_Tests;