+++ tests/tada_packages_tests.ads
@@ -4,4 +4,6 @@ package Tada_Packages_Tests is
    use Testy.Tests;
 
    procedure Test_Validate_Package_Names (T : in out Test_Context);
+
+   procedure Test_Validate_Package_Versions (T : in out Test_Context);
 end Tada_Packages_Tests;