+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+## 0.7.0
+
 * validate package versions: Semver + optional prerelease tag, example: `0.1.0-dev`
 * add `tada config` command: display configuration
 * support local and global toolchain configuration